Refactor components into common directory: Loading, NavigationBar. Refactor Hiscores component into Playground directory.
